refactor flashattention to ensure dtype consistency in gradient computation for Q and K tensors

+++ cs336_systems/flashattention.py
@@ -167,8 +167,8 @@ class FlashAttention(torch.autograd.Function):
         dV = einsum(P, dO, "... i j, ... i d -> ... j d")
         dP = einsum(dO.to(V.dtype), V, "... i d, ... j d -> ... i j")
         dS = P * (dP - D)
-        dQ = einsum(dS, K, "... a b, ... b c-> ... a c")/ctx.sqrt_d
-        dK = einsum(dS, Q, "... a b, ... a d -> ... b d")/ctx.sqrt_d
+        dQ = einsum(dS.to(K.dtype), K, "... a b, ... b c-> ... a c")/ctx.sqrt_d
+        dK = einsum(dS.to(Q.dtype), Q, "... a b, ... a d -> ... b d")/ctx.sqrt_d
 
         return dQ, dK, dV, None
